+++ src/main/java/org/folio/entlinks/domain/repository/AuthorityRepository.java
@@ -25,4 +25,6 @@ public interface AuthorityRepository extends JpaRepository<Authority, UUID>, Aut
   Optional<Authority> findByIdAndDeletedFalse(UUID id);
 
   List<Authority> findAllByIdInAndDeletedFalse(Collection<UUID> ids);
+
+  boolean existsAuthorityByAuthoritySourceFileId(UUID sourceFileId);
 }