+++ src/test/java/org/folio/entlinks/controller/converter/AuthoritySourceFileMapperTest.java
@@ -1,7 +1,6 @@
 package org.folio.entlinks.controller.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.folio.entlinks.domain.dto.AuthoritySourceFilePatchDto.SourceEnum.LOCAL;
 import static org.folio.support.base.TestConstants.INPUT_BASE_URL;
 import static org.folio.support.base.TestConstants.SOURCE_FILE_CODE;
 import static org.folio.support.base.TestConstants.SOURCE_FILE_NAME;
@@ -14,8 +13,10 @@ import org.apache.commons.lang3.StringUtils;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDto;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDtoCollection;
 import org.folio.entlinks.domain.dto.AuthoritySourceFilePatchDto;
+import org.folio.entlinks.domain.dto.AuthoritySourceFilePatchDtoHridManagement;
 import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDto;
 import org.folio.entlinks.domain.entity.AuthoritySourceFile;
+import org.folio.entlinks.domain.entity.AuthoritySourceFileCode;
 import org.folio.entlinks.domain.entity.AuthoritySourceFileSource;
 import org.folio.spring.testing.type.UnitTest;
 import org.jetbrains.annotations.NotNull;
@@ -75,23 +76,47 @@ class AuthoritySourceFileMapperTest {
 
   @Test
   void testPartialUpdate() {
-    AuthoritySourceFile sourceFile = createAuthoritySourceFile();
-
-    AuthoritySourceFilePatchDto patchDto = new AuthoritySourceFilePatchDto();
-    patchDto.setId(TEST_ID);
+    var sourceFile = createAuthoritySourceFile();
+    var patchDto = new AuthoritySourceFilePatchDto();
     patchDto.setName(UPDATED_NAME);
     patchDto.setType(UPDATED_TYPE);
     patchDto.setCodes(List.of(UPDATED_CODE));
     patchDto.setBaseUrl(UPDATED_BASE_URL);
-    patchDto.setSource(LOCAL);
+    patchDto.selectable(true);
+    patchDto.hridManagement(new AuthoritySourceFilePatchDtoHridManagement().startNumber(5));
 
-    AuthoritySourceFile updatedFile = mapper.partialUpdate(patchDto, sourceFile);
+    var updatedFile = mapper.partialUpdate(patchDto, sourceFile);
 
     assertThat(updatedFile).isNotNull();
-    assertThat(updatedFile.getId()).isEqualTo(patchDto.getId());
     assertThat(updatedFile.getName()).isEqualTo(patchDto.getName());
     assertThat(updatedFile.getType()).isEqualTo(patchDto.getType());
+    assertThat(updatedFile.getAuthoritySourceFileCodes().stream().map(AuthoritySourceFileCode::getCode).toList())
+        .isEqualTo(patchDto.getCodes());
     assertThat(updatedFile.getFullBaseUrl()).isEqualTo(patchDto.getBaseUrl());
+    assertThat(updatedFile.isSelectable()).isEqualTo(patchDto.getSelectable());
+    assertThat(updatedFile.getHridStartNumber()).isEqualTo(patchDto.getHridManagement().getStartNumber());
+    assertThat(updatedFile.getSource()).isEqualTo(sourceFile.getSource());
+    assertThat(updatedFile.getSequenceName()).isEqualTo(sourceFile.getSequenceName());
+  }
+
+  @Test
+  void testPartialUpdate_noCodesUpdate() {
+    var code = new AuthoritySourceFileCode();
+    code.setCode("a");
+    var codes = Set.of(code);
+    var sourceFile = createAuthoritySourceFile();
+    sourceFile.setAuthoritySourceFileCodes(codes);
+
+    var patchDto = new AuthoritySourceFilePatchDto();
+    patchDto.setName(UPDATED_NAME);
+    patchDto.setType(UPDATED_TYPE);
+
+    var updatedFile = mapper.partialUpdate(patchDto, sourceFile);
+
+    assertThat(updatedFile).isNotNull();
+    assertThat(updatedFile.getName()).isEqualTo(patchDto.getName());
+    assertThat(updatedFile.getType()).isEqualTo(patchDto.getType());
+    assertThat(updatedFile.getAuthoritySourceFileCodes()).isEqualTo(codes);
   }
 
   @Test
@@ -138,6 +163,8 @@ class AuthoritySourceFileMapperTest {
     sourceFile.setBaseUrl(INPUT_BASE_URL);
     sourceFile.setAuthoritySourceFileCodes(Set.of());
     sourceFile.setSource(AuthoritySourceFileSource.FOLIO);
+    sourceFile.setSelectable(false);
+    sourceFile.setHridStartNumber(1);
     return sourceFile;
   }